Add notify_incomplete_events option
Allows near-instant notifications while events are still occurring. If
set to No, notifications will only be sent once events complete.
However, enabling this setting disables image attachments and object
detection, because the maxscore frame image is not yet available. The
default value for this setting is No for this reason.