Remove notify_incomplete_events option; always use this behavior
Also re-enables image attachments and object detection when notifying
about incomplete events. It turns out the maxscore frame *is* available
before an event ends, contrary to the comments I wrote for the previous
commit. The only caveat is that the maxscore frame id is updated while
the event is occurring, so the one that it downloads may not end up
being the final maxscore frame. I think that is a fine compromise for
the time saved by notifying during running events.